requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.3
requests==2.28.2
Werkzeug==2.2.2
pytest==7.1.2
prometheus_flask_exporter==0.18.1
